pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "amipython"
version = "0.1.0"
description = "Python-to-Amiga game development toolchain"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
preview = [
    "pygame-ce>=2.4",
]

[project.scripts]
amipython = "amipython.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
amipython = ["c_runtime/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "docker: requires Docker with vbcc or ACE image",
]
